refactor step view
