+++ src/plugins/spark/components/BlockWidgetToolbar.vue
@@ -27,10 +27,22 @@ export default class BlockWidgetToolbar extends BlockCrudComponent {
       :config.sync="graphCfg"
     />
 
-    <q-item-section side>
+    <template #actions>
+      <ActionItem
+        v-if="hasGraph"
+        icon="mdi-chart-line"
+        label="Show graph"
+        @click="graphModalOpen = true"
+      />
+      <slot name="actions" />
+      <ActionItem icon="refresh" label="Refresh" @click="refreshBlock" />
+      <WidgetActions :crud="crud" no-rename />
+      <BlockActions :crud="crud" />
+    </template>
+
+    <!-- <q-item-section side>
       <q-btn-dropdown flat split icon="settings" @click="showDialog">
         <q-list dark bordered>
-          <!-- Global Actions -->
           <ActionItem
             v-if="hasGraph"
             icon="mdi-chart-line"
@@ -43,7 +55,7 @@ export default class BlockWidgetToolbar extends BlockCrudComponent {
           <BlockActions :crud="crud" />
         </q-list>
       </q-btn-dropdown>
-    </q-item-section>
+    </q-item-section> -->
   </WidgetToolbar>
 </template>
